Added a "Message this user" button to the user detail dialog
The button will either simply switch to the direct chat with the selected user,
or, if no such chat exists already, it will create a new room and invite the user
to it. It will properly switch the room message timeline view and room list tab view.
It will only appear if the focused room is not the direct chat with the user in question.